feat: test framework with contract events
