+++ test/framework/contracts.ts
@@ -7,6 +7,12 @@ interface DeployableContract<T> {
     deployed(): Promise<T>
 }
 
+/**
+ * Deploys a contract, that may or may not have constructor parameters.
+ *
+ * @param name the name of the contract in the Solidity file.
+ * @param args constract constructor arguments.
+ */
 export async function deployContract<T extends DeployableContract<T>>(
     name: string,
     ...args: Array<unknown>
@@ -17,12 +23,19 @@ export async function deployContract<T extends DeployableContract<T>>(
     return dao.deployed()
 }
 
+/**
+ * Executes the transaction and waits until it has processed before returning.
+ */
 export async function execute(
     transaction: Promise<ContractTransaction>
 ): Promise<ContractReceipt> {
     return (await transaction).wait()
 }
 
+/**
+ * Retrieves the signer found at the given index in the HardHat config,
+ * failing when not present.
+ */
 export async function signer(index: number): Promise<SignerWithAddress> {
     const signers = await ethers.getSigners()
     expect(signers.length).is.greaterThan(index)
