+++ test/framework/transaction.ts
@@ -5,6 +5,11 @@ import {ContractReceipt} from '@ethersproject/contracts/src.ts/index'
 // Transaction status code https://eips.ethereum.org/EIPS/eip-1066
 const SUCCESS = 1
 
+/**
+ * The expectation is successful transaction (with receipt).
+ *
+ * @param transaction waits for the receipt, verifying it is a success.
+ */
 export async function successfulTransaction(
     transaction: Promise<ContractTransaction>
 ): Promise<ContractReceipt> {
